+++ lib/components/Backdrop/Backdrop.d.ts
@@ -1,7 +1,8 @@
 import React from 'react';
 interface IBackdropProps {
+    target?: HTMLElement;
     onClick?: () => void;
     isTransparent?: boolean;
 }
-export declare const Backdrop: ({ isTransparent, onClick }: IBackdropProps) => React.ReactPortal;
+export declare const Backdrop: ({ target, isTransparent, onClick }: IBackdropProps) => React.ReactPortal;
 export {};